+++ src/components/quiz/Quiz.css
